Quizzical: some adjustments to the database
Found some design issues with the database (mostly with the `Quiz` and
`Question` tables) that had to be reworked. Discovered there might be some
confusion over what an `Entity` actually represents because I'm having to take
a 'category' and a 'name' -- might not be flexible enough?

I think I need to move away from Flask-SQLAlchemy. It's stupid to need a Flask
app to do anything with the database. Maybe I'll have a separate module for
command line maintanence of the database with SQLAlchemy so we don't have to
manually install test data.